Add TokenVerifier for non-OIDC-compliant JWTs
